Write to logs tab
